# openapi.yaml
openapi: 3.1.0
info:
  title: Maru API
  version: 0.1.0
  description: Versioned API for the Maru convention operating platform.
paths:
...
  /api/v1/me/notification-preferences/{organization_id}:
    get:
      operationId: communications_retrieve_my_preference
      parameters:
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreference'
          description: ''
    put:
      operationId: communications_update_my_preference
      parameters:
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationPreference'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreference'
          description: ''
  /api/v1/me/notifications:
    get:
      operationId: communications_list_my_notifications
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationMessage'
          description: ''
  /api/v1/me/notifications/{message_id}/read:
    post:
      operationId: communications_mark_my_notification_read
      parameters:
      - in: path
        name: message_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - me
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationMessage'
          description: ''
...
  /api/v1/organizations/{organization_id}/editions/{edition_id}/accreditation/me/credentials:
    get:
      operationId: accreditation_list_my_credentials
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credential'
          description: ''
...
  /api/v1/organizations/{organization_id}/editions/{edition_id}/credentials/{credential_id}/revoke:
    post:
      operationId: accreditation_revoke_credential
      parameters:
      - in: path
        name: credential_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialCommand'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
          description: ''
  /api/v1/organizations/{organization_id}/editions/{edition_id}/offline/conflicts:
    get:
      operationId: accreditation_list_offline_conflicts
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OfflineOperation'
          description: ''
  /api/v1/organizations/{organization_id}/editions/{edition_id}/offline/manifests:
    post:
      operationId: accreditation_generate_offline_manifest
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfflineManifest'
          description: ''
...
  /api/v1/organizations/{organization_id}/editions/{edition_id}/registrations/{registration_id}/credentials:
    post:
      operationId: accreditation_issue_credential
      parameters:
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: registration_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - organizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialCommand'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedCredential'
          description: ''
...
  /api/v1/public/organizations/{organization_id}/editions/{edition_id}/offline/devices/{device_code}/check-ins:
    post:
      operationId: accreditation_ingest_offline_check_in
      parameters:
      - in: path
        name: device_code
        schema:
          type: string
        required: true
      - in: path
        name: edition_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: organization_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - public
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfflineCheckIn'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfflineOperation'
          description: ''
...
components:
  schemas:
...
    ChannelEnum:
      enum:
      - email
      type: string
      description: '* `email` - Email'
...
    Credential:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        registration_id:
          type: string
          format: uuid
          readOnly: true
        public_id:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/CredentialStatusEnum'
          readOnly: true
        issue_sequence:
          type: integer
          readOnly: true
        label_snapshot:
          type: string
          readOnly: true
        issued_at:
          type: string
          format: date-time
          readOnly: true
        revoked_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        revocation_reason:
          type: string
          readOnly: true
      required:
      - id
      - issue_sequence
      - issued_at
      - label_snapshot
      - public_id
      - registration_id
      - revocation_reason
      - revoked_at
      - status
    CredentialCommand:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
      required:
      - reason
    CredentialStatusEnum:
      enum:
      - issued
      - replaced
      - revoked
      type: string
      description: |-
        * `issued` - Issued
        * `replaced` - Replaced
        * `revoked` - Revoked
...
    IssuedCredential:
      type: object
      properties:
        credential:
          $ref: '#/components/schemas/Credential'
        credential_token:
          type:
          - string
          - 'null'
      required:
      - credential
      - credential_token
...
    NotificationDelivery:
      type: object
      properties:
        channel:
          allOf:
          - $ref: '#/components/schemas/ChannelEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/NotificationDeliveryStatusEnum'
          readOnly: true
        attempt_count:
          type: integer
          readOnly: true
        safe_error_code:
          type: string
          readOnly: true
        last_attempt_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        delivered_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
      required:
      - attempt_count
      - channel
      - delivered_at
      - last_attempt_at
      - safe_error_code
      - status
    NotificationDeliveryStatusEnum:
      enum:
      - pending
      - succeeded
      - suppressed
      - permanent_failed
      type: string
      description: |-
        * `pending` - Pending
        * `succeeded` - Succeeded
        * `suppressed` - Suppressed by preference
        * `permanent_failed` - Permanent failure
    NotificationMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        organization_id:
          type: string
          format: uuid
          readOnly: true
        edition_id:
          type:
          - string
          - 'null'
          format: uuid
          readOnly: true
        message_type:
          type: string
          readOnly: true
        purpose:
          allOf:
          - $ref: '#/components/schemas/PurposeEnum'
          readOnly: true
        locale:
          type: string
          readOnly: true
        subject:
          type: string
          readOnly: true
        body:
          type: string
          readOnly: true
        action_path:
          type: string
          readOnly: true
        rendered_at:
          type: string
          format: date-time
          readOnly: true
        read_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        deliveries:
          type: array
          items:
            $ref: '#/components/schemas/NotificationDelivery'
      required:
      - action_path
      - body
      - deliveries
      - edition_id
      - id
      - locale
      - message_type
      - organization_id
      - purpose
      - read_at
      - rendered_at
      - subject
    NotificationPreference:
      type: object
      properties:
        organization_id:
          type: string
          format: uuid
          readOnly: true
        operational_email_enabled:
          type: boolean
        marketing_email_consent:
          type: boolean
        marketing_consent_version:
          type: string
          readOnly: true
        marketing_consented_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
      required:
      - marketing_consent_version
      - marketing_consented_at
      - organization_id
    OfflineCheckIn:
      type: object
      properties:
        operation_id:
          type: string
          format: uuid
        device_sequence:
          type: integer
          minimum: 1
        manifest_sequence:
          type: integer
          minimum: 1
        credential_token:
          type: string
          maxLength: 200
        occurred_at:
          type: string
          format: date-time
        signature:
          type: string
          maxLength: 128
      required:
      - credential_token
      - device_sequence
      - manifest_sequence
      - occurred_at
      - operation_id
      - signature
    OfflineManifest:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        sequence:
          type: integer
          readOnly: true
        valid_from:
          type: string
          format: date-time
          readOnly: true
        valid_until:
          type: string
          format: date-time
          readOnly: true
        generated_at:
          type: string
          format: date-time
          readOnly: true
        credential_count:
          type: integer
          readOnly: true
        payload:
          readOnly: true
        payload_digest:
          type: string
          readOnly: true
        signature:
          type: string
          readOnly: true
      required:
      - credential_count
      - generated_at
      - id
      - payload
      - payload_digest
      - sequence
      - signature
      - valid_from
      - valid_until
    OfflineOperation:
      type: object
      properties:
        operation_id:
          type: string
          format: uuid
          readOnly: true
        device_sequence:
          type: integer
          readOnly: true
        manifest_sequence:
          type: integer
          readOnly: true
        credential_public_id:
          type: string
          readOnly: true
        occurred_at:
          type: string
          format: date-time
          readOnly: true
        received_at:
          type: string
          format: date-time
          readOnly: true
        outcome:
          allOf:
          - $ref: '#/components/schemas/OfflineOperationOutcomeEnum'
          readOnly: true
        safe_result_code:
          type: string
          readOnly: true
      required:
      - credential_public_id
      - device_sequence
      - manifest_sequence
      - occurred_at
      - operation_id
      - outcome
      - received_at
      - safe_result_code
    OfflineOperationOutcomeEnum:
      enum:
      - applied
      - duplicate
      - conflict
      - rejected
      type: string
      description: |-
        * `applied` - Applied
        * `duplicate` - Duplicate
        * `conflict` - Conflict
        * `rejected` - Rejected
...
    PurposeEnum:
      enum:
      - operational
      - marketing
      - emergency
      type: string
      description: |-
        * `operational` - Operational service
        * `marketing` - Optional marketing
        * `emergency` - Emergency
...
    UpdateNotificationPreference:
      type: object
      properties:
        operational_email_enabled:
          type: boolean
        marketing_email_consent:
          type: boolean
        marketing_consent_version:
          type: string
          maxLength: 40
      required:
      - marketing_email_consent
      - operational_email_enabled
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid